코드리뷰 반영 - gradle custom task 파일로 분리
